pyproject.toml
[project]
name = "codaio-apicalls"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "tenacity>=8.2.2",
    "typer>=0.9.0",
    "beautifulsoup4>=4.12.3",
]
